back/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
accelerate
numpy
pillow
opencv-python-headless
python-multipart
starlette
